Add separate component to handle tooltips
